+++ scvi/model/base/__init__.py
@@ -2,6 +2,7 @@ from ._archesmixin import ArchesMixin
 from ._base_model import BaseModelClass
 from ._rnamixin import RNASeqMixin
 from ._training_mixin import UnsupervisedTrainingMixin
+from ._trainrunner import TrainRunner
 from ._vaemixin import VAEMixin
 
 __all__ = [
@@ -10,4 +11,5 @@ __all__ = [
     "RNASeqMixin",
     "VAEMixin",
     "UnsupervisedTrainingMixin",
+    "TrainRunner",
 ]
