+++ scvi/model/base/_training_mixin.py
@@ -3,9 +3,10 @@ from typing import Optional
 import numpy as np
 import torch
 
-from scvi.compose._training import TrainRunner
 from scvi.dataloaders import DataSplitter
 from scvi.lightning import TrainingPlan
+
+from ._trainrunner import TrainRunner
 
 
 class UnsupervisedTrainingMixin:
